Refactor pet forms to share component
